Rename test-without-watch script
